fix rc5 issues (#612)
* start branch


* fix serialize_flex_value error


* change fields to flexfloat


* migrate old graphics settings


* migrate orientation from compas to azimuth


* add migration tests


* fix ui unit hint for hp_power


* changed vat_production and vat_consumption from int-> float

* generated new docs

* fix limitation of pv_dc-production when chargelimit =0 (reported by @thesaxofonist)

* replaced NaN in battery overview with blancs

* update base images debian 9.3.0

* changed entrypoint to work with bashio

* changed build_miplib.sh to run in a batch

* changed run.sh to work with bashio and "use_self_compiled_miplib"

* added option "use_self_compiled_miplib" (default false)

* Version 2026.04.0.rc1

---------

+++ dao/prog/config/models/pricing.py
@@ -82,7 +82,7 @@ class PricingConfig(BaseModel):
             "x-validation-hint": "Dict with YYYY-MM-DD keys, float values (ex VAT)"
         }
     )
-    vat_consumption: dict[str, int] = Field(
+    vat_consumption: dict[str, float] = Field(
         alias="vat consumption",
         description="VAT percentage for consumption by date (YYYY-MM-DD -> %)",
         json_schema_extra={
@@ -92,7 +92,7 @@ class PricingConfig(BaseModel):
             "x-validation-hint": "Dict with YYYY-MM-DD keys, integer 0-100 values"
         }
     )
-    vat_production: dict[str, int] = Field(
+    vat_production: dict[str, float] = Field(
         alias="vat production",
         description="VAT percentage for production by date (YYYY-MM-DD -> %)",
         json_schema_extra={
